fix Unit is_on() for dimmable lights
I only have dimmable lights and in home assistant they are always
shown as but they are off in real.

The problem seems to be that _on is always True and only the dimmer
value is set to 0. The code seems to take care of that but the condition
was wrong, self._state.dimmer would always be false when the dimmer
is 0 thus it was going into the else branch which only return _on.

Just removing the condition makes it work because the dimmer value is
checked in the return line anyway.

Signed-off-by: Paul Holzinger <paul.holzinger@web.de>